changes for my linux server

+++ dbinit/initdb.sh
@@ -5,10 +5,15 @@
 EASYPHP=
 DB_USER=root
 DB_PASS=
-if [ "$OSTYPE" == "linux-gnu" ]; then
-	DB_PASS=murphy
-fi
 DB_HOST=localhost
+if [ "$OSTYPE" == "linux-gnu" ]; then
+    HOSTNAME=$(hostname)
+    if [ "$HOSTNAME" == "moon64" ]; then
+	DB_PASS=meister
+    else
+	DB_PASS=murphy
+    fi
+fi
 DATABASE=mysql
 
 SCRIPTDIR=$( dirname $0 )
@@ -19,30 +24,33 @@ if [ "$OS" == "Windows_NT" ]; then
 #	if [ -e "/c/xampp/mysql/bin/mysql.exe" ]; then
 #		MYSQL_EXE=/c/xampp/mysql/bin/mysql.exe
 #	else
-    	MYSQL_EXE=$(find $EASYPHP -type f -iname 'mysql.exe')
-
-	    if [ -z "$MYSQL_EXE" ]; then
-			echo "not found MySQL Executable, must quit!"
-			exit 3
-	    else
-			echo " found MySQL Executable in: $MYSQL_EXE"
-	    fi
-#	fi
+    MYSQL_EXE=$(find $EASYPHP -type f -iname 'mysql.exe')
+    
+    if [ -z "$MYSQL_EXE" ]; then
+	echo "not found MySQL Executable, must quit!"
+	exit 3
+    else
+	echo " found MySQL Executable in: $MYSQL_EXE"
+    fi
+    #	fi
 else
     # echo "This script need Windows NT environment to run, must quit!"
     # exit 1
     MYSQL_EXE=$(which mysql)
 fi
 
-export LIBMYSQL_ENABLE_CLEARTEXT_PLUGIN=1
-MYSQL_PATH=$(dirname $MYSQL_EXE)
-$MYSQL_PATH/mysqlcheck.exe --check-upgrade --all-databases  -u root
+# export LIBMYSQL_ENABLE_CLEARTEXT_PLUGIN=1
+# MYSQL_PATH=$(dirname $MYSQL_EXE)
+# $MYSQL_PATH/mysqlcheck.exe --check-upgrade --all-databases  -u root
 
 
 
 echo "clean DB"
-executeSql "drop user ''@'localhost';"
+if [ "$OS" == "Windows_NT" ]; then
+    executeSql "drop user ''@'localhost';"
+fi
 executeSql "drop user 't4user'@'%';"
+executeSql "drop user 't4user'@'localhost';"
 # executeSql "drop user 't3user'@'127.0.0.1';"
 # executeSql "drop user 't3user'@'localhost';"
 executeSql "flush privileges;"
@@ -53,9 +61,9 @@ executeSql "drop database web2;"
 # echo "mysql user: $MYSQLUSER"
 
 echo "init DB"
-executeSql "create user 't4user'@'%';"
+executeSql "create user 't4user'@'localhost';"
 executeSql "create database web2;"
-executeSql "GRANT ALL PRIVILEGES ON web2.* TO 't4user'@'%' WITH GRANT OPTION;"
+executeSql "GRANT ALL PRIVILEGES ON web2.* TO 't4user'@'localhost' WITH GRANT OPTION;"
 # executeSql "ALTER USER 't4user'@'%' IDENTIFIED BY PASSWORD('mPass#10');"
 # executeSql "SET old_passwords=0; SET PASSWORD FOR 't4user'@'%' = PASSWORD('passpass');"
 executeSql "UPDATE user SET password=password('passpass') WHERE user='t4user';"
